refactor: move ensure_accessibility_prompt to ffi/accessibility.rs
- Centralize accessibility permission check in FFI module
- main.rs reduced from 931 to 905 lines (-26 lines)
